updated services

+++ src/Lykke.Terminal.Domain/Trading/IOrderCalculator.cs
@@ -8,9 +8,8 @@ namespace Lykke.Terminal.Domain.Trading
 {
     public interface IOrderCalculator
     {
-        Task<double> CalculateProfitLossAsync(double openPrice, double closePrice, double volume, IAssetPair assetPair,
-            string baseAssetId);
+        Task<decimal> CalculateProfitLossAsync(decimal openPrice, decimal closePrice, double volume, IAssetPair assetPair, string baseAssetId);
 
-        Task<double> CalculateMarginUsageAsync(IEnumerable<IOrderBase> orders, ITradingAccount tradingAccount);
+        Task<decimal> CalculateMarginUsageAsync(IEnumerable<IOrderBase> orders, ITradingAccount tradingAccount);
     }
 }
