+++ src/Lykke.Terminal.Domain/Trading/ITradeExecutionService.cs
@@ -4,6 +4,6 @@ namespace Lykke.Terminal.Domain.Trading
 {
     public interface ITradeExecutionService
     {
-        Task<bool> CheckTradeRequestAsync(string orderId, string asset, double volume, double price);
+        Task<bool> CheckTradeRequestAsync(string orderId, string asset, double volume, decimal price);
     }
 }
